Pigale-1.3.24
previous commits are from https://sourceforge.net/p/pigale/code/
which offers a snapshot of the CVS commit history, but only up to 2009

+++ incl/TAXI/ndfs.h
@@ -110,7 +110,7 @@ class _Kuratowski : public _DFS
     Prop<int> ewidth;
     Prop<short> ecolor;
     Prop<short> vcolor;
-    Prop<long> elabel;
+    Prop<int> elabel;
     Prop<bool> mark;
     svector<tvertex> ancestor;
     svector<tvertex> ancestor_son;
